+++ apps/web/src/db/sessions/actions.ts
@@ -352,49 +352,6 @@ export async function updateEvents(session: Session, newEvent: EventMessage) {
     .where(eq(SessionsTable.short_id, session.short_id))
 }
 
-export async function addFeedback({
-  sessionId,
-  messageId,
-  feedback,
-  content,
-}: {
-  sessionId: string
-  messageId: string
-  feedback: 'good' | 'bad'
-  content?: string
-}) {
-  try {
-    const [{ messages_str }] = await db
-      .select()
-      .from(SessionsTable)
-      .where(eq(SessionsTable.short_id, sessionId))
-    if (!messages_str) {
-      return
-    }
-
-    const updatedMessage = JSON.parse(messages_str).map(
-      (message: ChatMessage) => {
-        if (message.id === messageId) {
-          message.feedback = feedback
-          message.feedback_content = content
-        }
-        return message
-      }
-    )
-
-    await db
-      .update(SessionsTable)
-      .set({
-        messages_str: JSON.stringify(updatedMessage),
-      })
-      .where(eq(SessionsTable.short_id, sessionId))
-  } catch (error: any) {
-    return {
-      error: error.message,
-    }
-  }
-}
-
 export async function getMonitoringData({
   appId,
   pageSize = 10,
